♻️ refactor(components): コンポーネント整理とコード重複削減
## 変更内容

### 未使用コンポーネントの削除 (7ファイル)
- ❌ Tooltip.tsx → Help.tsxに統合
- ❌ IconWithDot.tsx → 未使用のため削除
- ❌ Checkbox.tsx → 未使用のため削除
- ❌ Slider.tsx → RangeSlider.tsxに統合
- ❌ AuthWithSAML.tsx → Auth.tsxに統合
- ❌ AuthWithUserpool.tsx → Auth.tsxに統合
- ❌ AuthWithSamlOrUserpool.tsx → Auth.tsxに統合

### コンポーネント統合
- ✨ Auth系3コンポーネント → 1つのAuth.tsxに統合
  - mode propsで認証方法を切り替え ('userpool' | 'saml' | 'both')
  - コード重複を削減し、保守性を向上
- ✨ Help.tsx: Tooltip機能を直接統合
- ✨ PptxGenerationPage.tsx: Slider → RangeSliderに置き換え

## 効果
- 📉 コンポーネント数: -6ファイル
- ✅ 重複コード削減
- ✅ 依存関係の簡素化
- ✅ ビルド成功確認済み
