+++ packages/web/src/components/Help.tsx
@@ -1,7 +1,6 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 import { BaseProps } from '../@types/common';
 import { PiQuestionFill } from 'react-icons/pi';
-import Tooltip from './Tooltip';
 
 type Props = BaseProps & {
   message: string;
@@ -9,10 +8,27 @@ type Props = BaseProps & {
 };
 
 const Help: React.FC<Props> = (props) => {
+  const position = useMemo(() => {
+    switch (props.position) {
+      case 'left':
+        return 'right-0';
+      case 'right':
+        return 'left-0';
+      case 'center':
+        return 'left-1/2 -translate-x-1/2';
+    }
+  }, [props.position]);
+
   return (
-    <Tooltip {...props}>
+    <div className={`${props.className ?? ''} group relative`}>
+      <div
+        className={`invisible absolute ${position} -top-5 z-50 bg-transparent p-3 pl-5 pt-8 text-xs font-normal text-white opacity-0 transition group-hover:visible group-hover:opacity-100`}>
+        <div className="w-64 rounded border border-gray-400 bg-black/90 p-1">
+          {props.message}
+        </div>
+      </div>
       <PiQuestionFill />
-    </Tooltip>
+    </div>
   );
 };
